refactor: move button and dialog into component
